refactor(evalforge-skill-gate): split gate-run's helpers out by concern
gate-run.ts was 318 lines, a third of it helpers surrounding the flow. Now 196,
and what remains is the flow a reader comes to the file for.

  report.ts          describeError, fail, guardedCall, makeGateCommenter
  draft-tags.ts      isDraftTagActive — GitHub PR-state lookup
  apply-sync-plan.ts the CREATE/UPDATE/DELETE/DEFER apply loop
  run-eval.ts        startEvalRun, pollToCompletion

describeError was duplicated in gate-run and cleanup-run; report.ts is now the one
copy, so both runners share the failure primitives.

Behaviour unchanged — the code moved verbatim. The 70 tests drive runGate
end-to-end through mocked core, so they cover the extracted modules for real
rather than through doubles.
