+++ .github/actions/evalforge-skill-gate/src/utils/gate-run.ts
@@ -2,87 +2,20 @@ import * as core from '@actions/core';
 import * as github from '@actions/github';
 import { posix } from 'node:path';
 import {
-  EvalForgeClient, EvalRunTimeoutError, GATE_COMMENT_MARKER,
+  EvalForgeClient,
   collectSkillFiles, deriveTags, diffSyncPlan, draftTagFor, evalRunUrl, evaluateRunResult,
-  formatForeignDraftConflicts, formatGateResult, formatGateServiceError, formatGateTimeout,
+  formatForeignDraftConflicts, formatGateResult, formatGateServiceError,
   formatGuardFailure, formatNoGatedChanges, formatYamlErrors,
   getChangedFiles, getFirstCommitAuthorEmail, guardScenarios, isWixAuthorEmail,
-  listRemoteScenariosForGate, loadScenarios, makeCommenter, parseDraftTag, pollUntilDone,
-  remoteScenarioFiltersForGate, selectScenarios, semanticPlusDraftTags,
-  stripInactiveForeignDraftTags, touchedScenarioPaths,
-  type Commenter, type EvalRunStatus, type SyncAction,
+  listRemoteScenariosForGate, loadScenarios, remoteScenarioFiltersForGate, selectScenarios,
+  semanticPlusDraftTags, stripInactiveForeignDraftTags, touchedScenarioPaths,
 } from '@wix/evalforge-core';
-import { getGateConfig, BASE_WORKSPACE_SUBDIR, type GateConfig } from './config';
+import { getGateConfig, BASE_WORKSPACE_SUBDIR } from './config';
 import { workspaceRoot } from './workspace';
-
-function describeError(error: unknown): string {
-  return error instanceof Error ? error.message : String(error);
-}
-
-function fail(message: string, blocking: boolean): void {
-  if (blocking) core.setFailed(message);
-  else core.warning(message);
-}
-
-/** Runs an EvalForge call, reporting a user-facing comment and gate failure if it throws. */
-async function guardedCall<T>(
-  operation: () => Promise<T>,
-  userMessage: string,
-  comment: Commenter,
-  blocking: boolean,
-): Promise<T | undefined> {
-  try {
-    return await operation();
-  } catch (error) {
-    core.error(`${userMessage}: ${describeError(error)}`);
-    await comment(formatGateServiceError(userMessage, blocking));
-    fail(userMessage, blocking);
-    return undefined;
-  }
-}
-
-/** Timeout gets its own comment; anything else is a generic service failure. */
-async function pollToCompletion(
-  client: EvalForgeClient,
-  config: GateConfig,
-  runId: string,
-  runUrl: string,
-  comment: Commenter,
-): Promise<EvalRunStatus | undefined> {
-  try {
-    return await pollUntilDone(client, config.projectId, runId, {
-      log: core.info,
-      warn: core.warning,
-    });
-  } catch (error) {
-    if (error instanceof EvalRunTimeoutError) {
-      await comment(formatGateTimeout(runId, runUrl, config.blocking));
-      fail(error.message, config.blocking);
-      return undefined;
-    }
-    core.error(`Polling the eval run failed: ${describeError(error)}`);
-    await comment(formatGateServiceError('Polling the eval run failed', config.blocking));
-    fail('Polling the eval run failed', config.blocking);
-    return undefined;
-  }
-}
-
-async function isDraftTagActive(
-  octokit: ReturnType<typeof github.getOctokit>,
-  tag: string,
-): Promise<boolean> {
-  const draft = parseDraftTag(tag);
-  if (!draft) return true;
-  const [owner, repo] = draft.repo.split('/', 2);
-  if (!owner || !repo) return true;
-  try {
-    const pull = await octokit.rest.pulls.get({ owner, repo, pull_number: draft.prNumber });
-    return pull.data.state === 'open';
-  } catch (error) {
-    core.warning(`Could not resolve draft tag ${tag}: ${describeError(error)}`);
-    return true;
-  }
-}
+import { fail, guardedCall, makeGateCommenter } from './report';
+import { isDraftTagActive } from './draft-tags';
+import { applySyncPlan } from './apply-sync-plan';
+import { pollToCompletion, startEvalRun } from './run-eval';
 
 export async function runGate(): Promise<void> {
   const config = getGateConfig();
@@ -95,16 +28,7 @@ export async function runGate(): Promise<void> {
     return;
   }
 
-  const comment = makeCommenter(octokit, {
-    owner: config.owner,
-    repo: config.repo,
-    prNumber: config.prNumber,
-    marker: GATE_COMMENT_MARKER,
-  }, {
-    warn: core.warning,
-    writeSummary: async (body: string) => { await core.summary.addRaw(body).write(); },
-  });
-
+  const comment = makeGateCommenter(octokit, config);
   const workspace = workspaceRoot();
   const draftTag = draftTagFor(config.repoFullName, config.prNumber);
   core.info(
@@ -243,18 +167,7 @@ export async function runGate(): Promise<void> {
     return;
   }
 
-  const run = await guardedCall(
-    () => client.createAndRunEvalRun(config.projectId, {
-      name: `${config.repoFullName} PR #${config.prNumber} (${config.versionLabel})`,
-      description: `Skill gate run for PR #${config.prNumber}`,
-      projectId: config.projectId,
-      agentId: config.agentId,
-      scenarioIds: selection.ids,
-      capabilityIds: [config.capabilityId],
-      capabilityVersions: { [config.capabilityId]: config.versionLabel },
-    }),
-    'Could not start the eval run', comment, config.blocking,
-  );
+  const run = await startEvalRun(client, config, selection.ids, comment);
   if (!run) return;
 
   const runUrl = evalRunUrl(config.projectId, run.id);
@@ -280,39 +193,4 @@ export async function runGate(): Promise<void> {
   if (!verdict.passed) {
     fail(`Eval gate failed: ${verdict.reasons.join('; ')}`, config.blocking);
   }
-}
-
-/** Applies the plan, recording ids CREATE returns. Stops on failure: a half-synced set would
- * report on the wrong content. */
-async function applySyncPlan(
-  client: EvalForgeClient,
-  config: GateConfig,
-  actions: SyncAction[],
-  nameToId: Map<string, string>,
-  comment: Commenter,
-): Promise<boolean> {
-  for (const action of actions) {
-    try {
-      if (action.kind === 'CREATE') {
-        const created = await client.createTestScenario(config.projectId, action.body, action.tags);
-        nameToId.set(action.name, created.id);
-        core.info(`Created scenario ${action.name} (${created.id})`);
-      } else if (action.kind === 'UPDATE') {
-        await client.updateTestScenario(config.projectId, action.id, action.body, action.tags);
-        core.info(`Updated scenario ${action.name} (${action.id})`);
-      } else if (action.kind === 'DELETE') {
-        await client.deleteTestScenario(config.projectId, action.id);
-        nameToId.delete(action.name);
-        core.info(`Deleted draft scenario ${action.name} (${action.id})`);
-      } else {
-        core.info(`Deferring DELETE of "${action.name}" — handled at merge`);
-      }
-    } catch (error) {
-      core.error(`Sync action ${action.kind} for ${action.name} failed: ${describeError(error)}`);
-      await comment(formatGateServiceError(`Sync failed for "${action.name}"`, config.blocking));
-      fail(`Sync failed for ${action.name}`, config.blocking);
-      return false;
-    }
-  }
-  return true;
 }
